Conflicts:
	databases/RMG_database/kinetics_groups/H_Abstraction/dictionary.txt
	databases/RMG_database/kinetics_groups/H_Abstraction/rateLibrary.txt
	databases/RMG_database/thermo_groups/Abraham_Radical_Library.txt
	databases/RMG_database/thermo_groups/Abraham_Radical_Tree.txt
	source/RMG/Abraham.java
	source/RMG/RMG.java
	source/RMG/jing/chem/ChemGraph.java
	source/RMG/jing/chem/GATP_Solvation.java
	source/RMG/jing/chem/PrimaryAbrahamLibrary.java
	source/RMG/jing/chem/SolventData.java
	source/RMG/jing/chem/SolventLibrary.java
	source/RMG/jing/chem/ThermoGAGroupLibrary.java
	source/RMG/jing/rxn/Reaction.java
	source/RMG/jing/rxnSys/ReactionModelGenerator.java

These were resolved somewhat quickly, and a few things may be broken :-/

+++ source/RMG/jing/param/Global.java
@@ -2,7 +2,7 @@
 //
 //	RMG - Reaction Mechanism Generator
 //
-//	Copyright (c) 2002-2009 Prof. William H. Green (whgreen@mit.edu) and the
+//	Copyright (c) 2002-2011 Prof. William H. Green (whgreen@mit.edu) and the
 //	RMG Team (rmg_dev@mit.edu)
 //
 //	Permission is hereby granted, free of charge, to any person obtaining a
@@ -33,7 +33,7 @@ public class Global {
 	public static double RT_identifyReactedSites = 0;
 	public static double RT_reactChemGraph = 0;
 	public static double RT_findRateConstant = 0;
-	public static double tAtInitialization;
+	public static long tAtInitialization;
 	public static double makeSpecies = 0;
 	public static double checkReactionReverse = 0;
 	public static double makeTR = 0;
